fix: Handle fetch error gracefully and show error message
In the `Page()` function of `page.jsx`, added error handling for the `fetch` request to the API endpoint. If the network response is not ok, an error is thrown and caught. An error message is logged to the console and can be used to show an error message to the user.