fix: Add defaultChain to Web3Modal component
This commit fixes a bug in the `provider.jsx` file, where the `defaultChain` prop was not being passed to the `Web3Modal` component.

The `defaultChain` prop is now added to the `Web3Modal` component, which will provide the default chain value for the modal.

The `helachain` variable is now included in the `initializeApp` function and passed as the `defaultChain` prop to the `Web3Modal` component.