Fix: Fix logout functionality in navbar component
This commit fixes the issue where the logout functionality in the navbar component was not working properly. This was due to a missing disconnectAsync function call. Now, the user will be properly disconnected and logged out when clicking the logout button.

Note: This commit doesn't add any new features, it only fixes a bug in the existing code.