feat: Add refreshMetadata function
Added a new function `refreshMetadata` to the Auction component that allows the user to refresh metadata for a specific NFT. The function makes a POST request to the server with the collection address and token ID, and updates the state `refreshMetada` to trigger a re-fetch of the auction data.

The `refreshMetadata` function is called when the "Refresh Metadata" button is clicked in the Slideshow component.

This change adds new functionality to the application and improves the user experience by allowing the user to update the metadata for an NFT in the auction.